requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0
openai>=1.3.0
reportlab>=4.0.4
Pillow>=10.0.0
httpx[socks]>=0.23.0
